Execute subprocess in a thread
Currently when we execute the script in the button callback, the UI
freezes and the shell creates an "application not responding" dialog on
the app. Avoid that by executing the (potentially long running) script
in a separate thread, and updating the UI in a callback.

Use the Gio API for this, so that the callback will happen in the main
thread, which is a requirement of GTK.